.ipynb_checkpoints/requirements-checkpoint.txt
langchain
langchain-google-genai
tavily-python
python-dotenv
